--- src/constants/screenNames.ts
@@ -1,5 +1,0 @@
-export const SCREEN_NAME = {
-  WELCOME_SCREEN: 'WelcomeScreen',
-  HOME_SCREEN: 'HomeScreen',
-  DETAIL_SCREEN: 'DetailScreen'
-};
