+++ src/screens/HomeScreen.js
@@ -1,4 +1,4 @@
-import {View, Text} from 'react-native';
+import { View, Text, StyleSheet } from 'react-native';
 import React from 'react';
 
 const HomeScreen = () => {
@@ -8,5 +8,7 @@ const HomeScreen = () => {
     </View>
   );
 };
+
+const styles = StyleSheet.create({});
 
 export default HomeScreen;
