--- src/screens/index.ts
@@ -1,3 +1,0 @@
-export { default as WelcomeScreen } from './welcome';
-export { default as HomeScreen } from './home';
-export { default as DetailScreen } from './detail';
