--- src/screens/welcome/styles.ts
@@ -1,5 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({});
-
-export default styles;
