+++ tsconfig.json
@@ -9,16 +9,9 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "noEmit": true,
-    "baseUrl": ".",
     "jsx": "react",
     "skipLibCheck": false,
-    "paths": {
-      "screens": ["./src/screens/*"],
-      "constants": ["./src/constants/*"],
-      "router": ["./src/router/*"],
-      "assets": ["./src/assets/*"],
-      "components": ["./src/components/*"]
-    }
+   
   },
   "exclude": ["node_modules", "babel.config.js"]
 }
